+++ iosApp/iosApp/UI/ContentView.swift
@@ -4,13 +4,13 @@ import Combine
 
 ///SwiftUI TabView
 struct ContentView: View {
-    @StateObject var viewModel = DevFestViewModel()
+    @StateObject var viewModel = BaseViewModel()
     
     @State private var selection = 0
     
     var body: some View {
         TabView(selection: $selection) {
-            AgendaView(viewModel: viewModel)
+            AgendaView()
                 .tabItem {
                     VStack {
                         Image(systemName: "calendar")
@@ -18,21 +18,29 @@ struct ContentView: View {
                     }
                 }.tag(0)
             
-            VenueView(viewModel: viewModel)
+            SpeakersView()
+                .tabItem {
+                    VStack {
+                        Image(systemName: "person.3.fill")
+                        Text(L10n.screenSpeakers)
+                    }
+                }.tag(1)
+            
+            VenueView()
                 .tabItem {
                     VStack {
                         Image(systemName: "location.circle.fill")
                         Text(L10n.screenVenue)
                     }
-                }.tag(1)
+                }.tag(2)
             
-            AboutView(viewModel: viewModel)
+            AboutView()
                 .tabItem {
                     VStack {
                         Image(systemName: "info.circle")
                         Text(L10n.screenAbout)
                     }
-                }.tag(2)
+                }.tag(3)
         }
         .onAppear{
             // correct the transparency bug for Tab bars